chore: Header 컴포넌트에  사이드바 버튼 수정
